day 9

+++ day9/src/file.c
@@ -1,6 +1,6 @@
 #include "../include/day9.h"
 
-int open_file(char* filename) 
+long long open_file(char* filename) 
 {
     FILE *fd = fopen(filename, "r");
     if (fd == NULL) {
