+++ day9/src/main.c
@@ -3,8 +3,8 @@
 int main (int argc, char** argv)
 {
     if (argc == 2) {
-        int value = open_file(argv[argc - 1]);
-        printf("answer: %i\n", value);
+        long long value = open_file(argv[argc - 1]);
+        printf("answer: %lld\n", value);
         return EXIT_SUCCESS;
     }
     return EXIT_FAILURE;
